Update Auth pages components
